making file executable

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 
 /**
- * print_dlistint - prints all the elements of a dlistint_t list.
+ * print_dlistint - prints all the elements of a list.
  * @h: first node in the list
  *
  * Description: doubly linked list node structure
@@ -12,7 +12,7 @@
 size_t print_dlistint(const dlistint_t *h)
 {
 	const dlistint_t *current = h;
-	size_t count = 0;
+	int count = 0;
 	while (current != NULL)
 	{
 		printf("%d\n", current->n);
